Refactor: move habits progress bar
