+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added:
+- Chart for habit completion rate over time
+
+## [0.8.204] - 2022-12-07
+### Added:
 - Progress bar for habit progress for the current day
 
 ## [0.8.203] - 2022-12-05
